requirements.txt
pytesseract
opencv-python
Pillow
gspread
oauth2client
streamlit
